+++ src/main/java/dealership/vehicles/ElectricCar.java
@@ -5,7 +5,7 @@ import dealership.components.Engine;
 
 public class ElectricCar extends Vehicle {
 
-    public ElectricCar(String make, int price, Engine engine) {
+    public ElectricCar(String make, double price, Engine engine) {
         super(make, price, engine);
     }
 }
