+++ src/main/java/dealership/vehicles/Hybrid.java
@@ -5,7 +5,7 @@ import dealership.components.Engine;
 
 public class Hybrid extends Vehicle {
 
-    public Hybrid(String make, int price, Engine engine) {
+    public Hybrid(String make, double price, Engine engine) {
         super(make, price, engine);
     }
 }
