+++ src/main/java/dealership/vehicles/Petrol.java
@@ -5,7 +5,7 @@ import dealership.components.Engine;
 
 public class Petrol extends Vehicle {
 
-    public Petrol(String make, int price, Engine engine) {
+    public Petrol(String make, double price, Engine engine) {
         super(make, price, engine);
     }
 }
